+++ back-end/src/main/kotlin/com/apexdevs/backend/ape/entity/workflow/Data.kt
@@ -11,7 +11,6 @@ import org.json.JSONObject
  * @param taxonomyRoots a map that uses the taxonomy dimensions to format the data e.g. ["type"->"Image","Format"->"PNG"]
  */
 data class Data(val taxonomyRoots: MutableMap<String, String>) {
-
     fun toJSONObject(): JSONObject {
         val jsonObject = JSONObject()
         taxonomyRoots.forEach { entry ->
@@ -19,4 +18,21 @@ data class Data(val taxonomyRoots: MutableMap<String, String>) {
         }
         return jsonObject
     }
+}
+
+/**
+ * Construct a list of Data objects from JSON data.
+ * @param json The JSON data.
+ * @return A list of Data objects constructed from the given JSON data.
+ */
+fun dataListFromJSON(json: ArrayList<LinkedHashMap<String, ArrayList<String>>>): List<Data> {
+    val result = mutableListOf<Data>()
+    json.forEach { linkedHashMap ->
+        val mutableMap: MutableMap<String, String> = LinkedHashMap()
+        linkedHashMap.forEach { entry ->
+            mutableMap[entry.key] = entry.value[0]
+        }
+        result.add(Data(mutableMap))
+    }
+    return result
 }