+++ back-end/src/main/kotlin/com/apexdevs/backend/persistence/exception/DomainVerificationNotFoundException.kt
@@ -1,0 +1,5 @@
+package com.apexdevs.backend.persistence.exception
+
+import org.bson.types.ObjectId
+
+class DomainVerificationNotFoundException(val from: Any, val domainId: ObjectId, message: String) : RuntimeException(message)