+++ front-end/src/components/Domain/DomainEdit/DomainEdit.tsx
@@ -11,7 +11,7 @@ import { Button, Col, Form, Input, message, Popconfirm, Row, Select, Space, Uplo
 import { DownloadOutlined, InfoOutlined, UploadOutlined } from '@ant-design/icons';
 import { UploadFile } from 'antd/lib/upload/interface';
 import { validateJSON, validateOWL, onFileChange, ReadMultipleFileContents, RMFCInput } from '@helpers/Files';
-import Domain, { Topic, Visibility } from '@models/Domain';
+import { DomainDetails, Topic, Visibility } from '@models/Domain';
 import { constraintsModal, ontologyModal, runConfigModal, toolAnnotationsModal } from '@components/Domain/Domain';
 import styles from './DomainEdit.module.less';
 
@@ -22,11 +22,13 @@ const { Option } = Select;
  */
 interface IProps {
   /** The domain to edit */
-  domain: Domain,
+  domain: DomainDetails,
   /** The topics that are available to choose from */
   topics: Topic[],
   /** Router used for redirecting after save or cancel */
   router: NextRouter,
+  /** Callback function called when the domain is updated. */
+  onUpdated?: () => void,
 }
 
 /**
@@ -198,7 +200,7 @@ class DomainEdit extends React.Component<IProps, IState> {
    */
   handleSubmit = (values: any) => {
     // Add any changed values to the payload
-    const { domain } = this.props;
+    const { domain, onUpdated } = this.props;
     const { owlFiles, toolsAnnotationsFiles, runConfigFiles, constraintsFiles } = this.state;
     // Add topics of they are changed
     const { appliedTopics, topicsChanged } = this.state;
@@ -262,6 +264,9 @@ class DomainEdit extends React.Component<IProps, IState> {
             return;
           }
           message.success('Domain has been updated');
+          if (onUpdated !== null) {
+            onUpdated();
+          }
         });
     });
   };
@@ -307,7 +312,7 @@ class DomainEdit extends React.Component<IProps, IState> {
           wrapperCol={{ span: 14 }}
           initialValues={{
             ...domain,
-            strictToolsAnnotations: domain.strictToolsAnnotations.toString(),
+            strictToolsAnnotations: domain.strictToolAnnotations.toString(),
           }}
           onFinish={this.handleSubmit}
           className={styles['Domain-edit']}